+++ packages/addons/dms/src/index.ts
@@ -13,3 +13,6 @@ export * from './client';
 // Service
 export * from './dms.service';
 export * from './dms-sync.service';
+
+// Module
+export * from './dms.module';